requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.35.3
pydantic==2.7.4
pydantic-core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
typing-extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
